requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bs4==0.0.1
cachetools==5.3.2
camelot-py==0.11.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.20
click==8.1.7
coloredlogs==15.0.1
comm==0.2.0
contourpy==1.2.0
cryptography==42.0.1
cycler==0.12.1
dataclasses-json==0.6.3
debugpy==1.8.0
decorator==5.1.1
Deprecated==1.2.14
distro==1.8.0
dnspython==2.4.2
easyocr==1.7.1
Elixir==0.7.1
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.105.0
filelock==3.13.1
flatbuffers==23.5.26
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
ghostscript==0.7
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
googleapis-common-protos==1.62.0
greenlet==3.0.2
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
idna==3.6
imageio==2.33.1
importlib-metadata==6.11.0
importlib-resources==6.1.1
ipykernel==6.27.1
ipython==8.19.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.5.1
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.352
langchain-community==0.0.6
langchain-core==0.1.3
langsmith==0.0.75
lazy_loader==0.3
llama-index==0.9.29
loguru==0.7.2
Markdown==3.5.2
markdown-it-py==3.0.0
marko==2.0.2
MarkupSafe==2.1.4
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdformat==0.7.17
mdurl==0.1.2
mmh3==4.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
neo4j==5.15.0
neomodel==5.2.1
nest-asyncio==1.5.8
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8.1
numpy==1.26.2
oauthlib==3.2.2
onnxruntime==1.16.3
openai==1.5.0
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation==0.42b0
opentelemetry-instrumentation-asgi==0.42b0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
opentelemetry-util-http==0.42b0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
parso==0.8.3
pbr==6.0.0
pdf2image==1.17.0
pdfminer.six==20231228
pexpect==4.9.0
Pillow==10.1.0
pinecone-client==2.2.4
platformdirs==4.1.0
posthog==3.1.0
prompt-toolkit==3.0.43
protobuf==4.25.1
psutil==5.9.7
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.17.2
pyparsing==3.1.1
pypdf==4.0.0
PyPika==0.48.9
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
QtPy==2.4.1
rapidfuzz==3.5.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.11.4
shapely==2.0.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.4
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tabulate==0.9.0
Tempita==0.5.2
tenacity==8.2.3
thefuzz==0.20.0
tifffile==2023.12.9
tiktoken==0.5.2
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
transformers==4.37.1
typer==0.9.0
types-protobuf==4.24.0.4
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
watchfiles==0.21.0
wcwidth==0.2.12
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
xlrd==0.7.1
xlwt==0.7.2
yarl==1.9.4
zipp==3.17.0
